+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ========
+## [0.1.3]
+- Allow custom `libraryIcon` URL to be used in wepub-viewer uplink (navigation).
 
 ## [0.1.2]